[thermalctld] Fix fan status issue in 201911 (#96)
Previous PR #92 has been merged to 201911. The previous fix used a method FanStatus.is_ok, however, the implementation of this method is different between master and 201911. So we need an extra enhance for this issue.

fan status should be true only if FanStatus.is_ok() and fan_status is true.