[xcvrd] Implement to read sfp info from eeprom several times while detected sfp inserted
It is known that some sfp modules will need more time to power up its
data path and sometimes it caused that the application takes the
incorrect or uninitialized data to display.

The original way is to read eeprom as soon as the sfp is inserted. And
the latest modification in PR(https://github.com/Azure/sonic-platform-daemons/pull/201)
is to add a 2 seconds delay after sfp is detected.

Our modification is to keep the original way which will read eeprom as
soon as the sfp is detected. It is expected that the eeprom data in
some sfp modules which have shorter warm up time can be obtained
immediately. For the other modules which has longer init time, xcvrd
will have a second try to read eeprom when sfp is inserted for 2 seconds
and the third try will be applied when sfp is inserted for 5 seconds.
Record to syslog when the checksum of sfp info read from eeprom is
different from the last try.