Modular chassis: Psud set master led on first run (#206)
On first run, the PSU master-led on supervisor of modular-chassis is not set.

Description
The PSU master-led on modular chassis indicates the power-buget status. Similar to individual PSU leds that are set on first_run, the master-led also will be set to based on initial status (RED or GREEN with LED status ON).

Motivation and Context
Without the forced setting of first_run, the master-led status will default to per vendor default. In some cases, this could be amber blinking etc