[ycabled] add capability to enable/disable telemetry (#279)
This PR provides a capability to sonic-utilities CLI to enable/disable telemetry for ycabled.
Basically there is a periodic loop for ycabled which posts telemetry data for that configured interval of time(currently 60 sec). This PR diables this data posting, and does not call platform API calls for ycable.
This PR is required for the initiative of getting some failover/switchover not get interfered because of sonic-telemetry API calls.
The CLI for enabling/disabling telemetry is
config muxcable telemetry enable/disable

Description
Motivation and Context
How Has This Been Tested?
UT and deploying changes on Arista testbed

Signed-off-by: vaibhav-dahiya <vdahiya@microsoft.com>