chassisd: Fix crash on exit on linecard
Set the `config_manager` variable to `None` if we are running on a
linecard and thus don't need to set up the config manager.

During cleanup, the chassid service tries to clean up the
`config_manager`, but the `config_manager` variable is only ever
initialized if we are on the supervisor. Thus, checking if it is
`None` is insufficient because this results in an `UnboundLocalError`
that prevents the cleanup from succeeding on a linecard.