Update sonic-ycabled Python package build for Bookworm
The build infra for Python packages in Bookworm has changed.
Specifically, by default, a sdist (source) tarball is built first, and
then the wheel is built. Because of this, the proto file needs to be
included in the sdist tarball, and the Python bindings need to be
generated at a different stage.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>