Credo AEC 400G firmware v4.1 needs to set low power mode when change to DPDeinit state
Description
Credo AEC 400G needs to set low power mode when changing to DPDeinit state
on firmware v4.1. After https://github.com/sonic-net/sonic-platform-daemons/pull/254
is merged, low power mode will not be set when changing to DPDeinit state.

Motivation and Context
This patch adds the code to set low power mode only when changing to
DPDeinit state when the transceiver's vendor is 'Credo' and 'host_assign'
is '1'(Which means the application mode only has 1 logical port on the transceiver).
With this patch, the Credo AEC 400G can work correctly.