[xcvrd] DPB support on platforms with CmisManagerTask enabled (#500)
* fix(xcvrd): fix DPB support with CMIS transceivers

CmisManagerTask's `port_dict` and `port_mapping` must be updated
according to the port add/remove events.

Before this commit, `port_mapping` is only intialized when
CmisManagerTask is initialized and not updated after that,
which was causing KeyError exception when DBP is used.
(https://github.com/sonic-net/sonic-buildimage/issues/18893)

This commit removes the `port_mapping` field from CmisManagerTask as
`port_mapping` was used just for storing `asic_id` information
and that can be simply done by `port_dict` instead.

Also, this commit updates `port_dict` accoding to the port add/remove
events to support DPB.

Signed-off-by: Wataru Ishida <wataru.ishid@gmail.com>

* fix(xcvr): improve coverage

Signed-off-by: Wataru Ishida <wataru.ishid@gmail.com>

* fix(xcvrd): fix PORT_DEL handling due to xcvr plug-out

Signed-off-by: Wataru Ishida <wataru.ishid@gmail.com>

* refactor(xcvrd): remove sfputilhelper

it is not used.

Signed-off-by: Wataru Ishida <wataru.ishid@gmail.com>

---------

Signed-off-by: Wataru Ishida <wataru.ishid@gmail.com>