Refactor ledd daemon & fix high CPU usage due to unexpected socket close (#548)
Description
[ledd] Refactor ledd daemon and fix high CPU usage due to unexpected redis DB socket close

As part of the refactor, the LED port color is changed to oper UP only if all subports of a front panel ports are UP.
If any subport is oper DOWN, the LED port color is changed to oper DOWN

Motivation and Context
For unkown reason if redis closes the socket, then ledd daemon is continuously busy waiting for socket event and failing all the time. This is resulting in high CPI

The change in Port LED behavior is necessary to prevent situation in the old behavior where even if some subports are oper down, the front panel port LED should not glow green.

How Has This Been Tested?
Close the socket abruptly and see ledd daemon is respawned so that new socket connection is established to redis DB without causing high CPU.

Test this on Arista 7060X6 with 8x100G mode and 2x400G mode