[Smartswitch][Chassisd] Fix for initialization DPU admin state post startup and pre shutdown (#629)
This change is so that the pre shutdown and post startup are handled even during the dark mode initialization for the DPUs.
Initial DPU state changes are handled by submit_dpu_callback so we need the pre/post execution for initialization

Motivation and Context
During initialization, the admin state change for the DPU is handled by submit_dpu_callback function, so we need the module_pre_shutdown and module_post_startup calls in submit_dpu_callback as well.