[Smartswitch][Chassisd] Fix for initialization DPU admin state post startup and pre shutdown (#634)
<!-- Provide a general summary of your changes in the Title above -->

#### Description
<!--
 Describe your changes in detail
-->

This change is so that the pre shutdown and post startup are handled even during the dark mode initialization for the DPUs.
Initial DPU state changes are handled by submit_dpu_callback so we need the pre/post execution for initialization

#### Motivation and Context
<!--
 Why is this change required? What problem does it solve?
 If this pull request closes/resolves an open Issue, make sure you
 include the text "fixes #xxxx", "closes #xxxx" or "resolves #xxxx" here
-->
During initialization, the admin state change for the DPU is handled by `submit_dpu_callback` function, so we need the `module_pre_shutdown` and `module_post_startup` calls in `submit_dpu_callback` as well.

#### How Has This Been Tested?
<!--
 Please describe in detail how you tested your changes.
 Include details of your testing environment, and the tests you ran to
 see how your change affects other areas of the code, etc.
-->
Unit test and manual tests

#### Additional Information (Optional)