Chassisd does not wait for the execution to complete for previous admin state change requests. Fixed issue #22138 (#607)
Fixed issue #22138
Chassisd does not wait for the execution to complete for previous admin state change requests. Added support to solve this.

The corresponding sonic-utilities PR is: sonic-net/sonic-utilities#3937

Motivation and Context
Added a state to indicate the progress of an admin_state transition period (window) of every module in config_db.
When in progress do not allow another admin_state change for that module.
Mark the beginning of state transition in the CLI itself. Added a timeout also for error cases.
Mark the end of state transition in chassisd