update phy entity tables to account for modified psu key naming (#640)
Description
Modify every instance of PSU parent naming to attempt to call the get_name API from the PSU/parent object, falling back to the prior implemented default behavior in the case that the API is not yet implemented by the vendor.

Motivation and Context
Issue raised: #638, the key changes that took place previously on PSUD missing updating their associated phy entity tables in thermalctld. This PR makes the changes to account for this change.