[202505] [xcvrd] Improve logging in case of unable to find appropriate match for optic SI settings (#672) (#673)
202505 cherry-pick for #672

Description
This PR improves the logging in the optics SI parser to provide better diagnostic information when unable to find an appropriate match for optic SI settings for a physical port.

Motivation and Context
The previous log message was an error which is misleading since there can be genuine cases wherein optics SI settings are not present for a module.
Hence, changing the log level to INFO and adding more debug data.

How Has This Been Tested?
Verified

Additional Information (Optional)
MSFT ADO - 34731836

Signed-off-by: Mihir Patel <patelmi@microsoft.com>