[xcvrd] Add new thread to poll temperatures more aggressively (#685)
* xcvrd: Add new thread to poll temperatures aggressively

For high power modules it is important to query the temperature often as they can
heat up quickly and the cooling algorithm needs to react quickly to avoid damaging
the transceiver modules.

Introducing a new DomThermalUpdateInfoTask thread which is solely polling the
temperature information out of the xcvrs and publishing these in a new
TRANSCEIVER_DOM_TEMPERATURE table in STATE_DB.

This is mostly additional code being added.
To maximize code reuse, the existing DomUpdateInfoTask has been refactored so that
it now share a base with the newly introduced DomThermalInfoUpdateTask.

Benchmarking shows that polling the temperature out of a xcvrand publishing it inr
the databasse takes around 0.0016 for 800G optics.
Considering a worst case of 0.002 and considering a device with 64 ports, this
amounts to around 0.128s for every loop.
With this loop running every 5s this means this represent a ~2.5% of busyness time.

* xcvrd: Address review comments

Make the new DomThermalInfoUpdateTask optional. A program argument is
now necessary for xcvrd which can be set via the `pmon_daemon_control.json`

The new `--dom_temperature_poll_interval` option takes an update
interval in seconds.

Address misc comments

* xcvrd: fix test issue and address more comments

* xcvrd: improve test coverage

* fix invalid reference to helper_logger